requirements.txt
numpy
Pillow